+++ jogl-samples/src/tests/gl_320/primitive/Gl_320_primitive_sprite.java
@@ -35,7 +35,7 @@ public class Gl_320_primitive_sprite extends Test {
     }
 
     public Gl_320_primitive_sprite() {
-        super("gl-320-primitive-sprite", Profile.CORE, 3, 2, new Vec2((float) Math.PI * 0.2f));
+        super("gl-320-primitive-sprite", Profile.CORE, 3, 2, new Vec2(Math.PI * 0.2f));
     }
 
     private final String SHADERS_SOURCE = "primitive-sprite";
